+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -31,8 +31,9 @@ class UserServiceTest {
     @Test
     void testChangePassword() {
         final var dataSource = DataSourceConfig.getInstance();
-        final var userHistoryDao = new UserHistoryDao(dataSource);
-        final var userService = new UserService(dataSource, userDao, userHistoryDao);
+        final var userHistoryDao = new UserHistoryDao(jdbcTemplate);
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        final var userService = new TxUserService(dataSource, appUserService);
 
         final var newPassword = "qqqqq";
         final var createBy = "gugu";
@@ -48,7 +49,10 @@ class UserServiceTest {
         final var dataSource = DataSourceConfig.getInstance();
         // 트랜잭션 롤백 테스트를 위해 mock으로 교체
         final var userHistoryDao = new MockUserHistoryDao(jdbcTemplate);
-        final var userService = new UserService(dataSource, userDao, userHistoryDao);
+        // 애플리케이션 서비스
+        final var appUserService = new AppUserService(userDao, userHistoryDao);
+        // 트랜잭션 서비스 추상화
+        final var userService = new TxUserService(dataSource, appUserService);
 
         final var newPassword = "newPassword";
         final var createBy = "gugu";
